refactor(plugins): move PluginDashboards under plugins/admin
