+++ public/app/features/plugins/admin/components/PluginDashboards.tsx
