+++ public/app/features/plugins/admin/components/PluginDetailsBody.tsx
@@ -8,7 +8,7 @@ import { CatalogPlugin, PluginTabIds } from '../types';
 import { VersionList } from '../components/VersionList';
 import { usePluginConfig } from '../hooks/usePluginConfig';
 import { AppConfigCtrlWrapper } from './AppConfigWrapper';
-import { PluginDashboards } from '../../PluginDashboards';
+import { PluginDashboards } from './PluginDashboards';
 
 type Props = {
   plugin: CatalogPlugin;
